+++ NEWS.md
@@ -1,5 +1,10 @@
 # ggThemeAssist 0.1.1.9000
 
+## New features
+* Set legend position by click
+
+
+## Bugfixes
 * Plot height is now a relative value to improve visibility in combinations with lower screen resolutions. Closes #32.
 
 # ggThemeAssist 0.1.0